// src/app/modules/kentico/models/ilteam.model.ts
import { ContentItem, Elements } from 'kentico-cloud-delivery';
export class Team extends ContentItem {

  images_header: Elements.AssetsElement;
  title: Elements.RichTextElement;
  subtitle: Elements.RichTextElement;
  images_body: Elements.AssetsElement;
  staff: Elements.LinkedItemsElement;
}